Updated sim.py
working with algo.py. About 44 seconds for 2048 episodes with the cpu, 66 seconds with cuda. Now need to see if the results are good. With cpu, it reached a solution after 30,000 episodes with a reward of -39.7, which does seem higher than Natural Decay. Examine the results of the algo, and see if any adjustments need to be made.